requirements.txt
beautifulsoup4==4.14.2
certifi==2025.11.12
chardet==5.2.0
charset-normalizer==3.4.4
cssselect==1.3.0
idna==3.11
lxml==6.0.2
lxml_html_clean==0.4.3
psycopg2-binary==2.9.11
readability-lxml==0.8.4.1
requests==2.32.5
soupsieve==2.8
typing_extensions==4.15.0
urllib3==2.5.0
